Endpoint Modification d'un équipement et logiques dans les couches suivantes

+++ src/main/java/fr/karspa/hiker_thinker/dtos/EquipmentDTO.java
@@ -15,7 +15,7 @@ public class EquipmentDTO {
     private String description;
     private String brand;
 
-    private Float weight;
+    private Integer weight;
 
     private String categoryName;
